pyproject.toml
[project]
name = "projetointerdisciplinar"
version = "0.1.0"
description = "Backend for the learning app developed for Projeto Interdisciplinar"
authors = [
    {name = "jlxmns",email = "jlxmns@pm.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)"
]

[tool.ruff]
fix = true
line-length = 88

exclude = ["**/migrations"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
